% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_name.R
\name{change_name}
\alias{change_name}
\title{Change R package name}
\usage{
change_name(new.name = NULL, gh.user = NULL)
}
\arguments{
\item{new.name}{character. New name of the package.}

\item{gh.user}{character. GitHub user or organisation name where the new repository
is hosted.}
}
\value{
Modified file names and contents.
}
\description{
Change R package name from 'templateRpackage' to the chosen new name, by replacing
all mentions to 'templateRpackage' within files and file names.
Note this function is potentially dangerous, as it changes file names and contents,
and deleted the entire 'man' folder.
Only run this function just after creating the new repository from the template.
}
\details{
Inspired on the \href{https://cran.r-project.org/package=changer}{changer}
package by Jouni Helske.
}
