% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_array.R
\name{sum_array}
\alias{sum_array}
\title{Sum a numeric vector using optimized C++ code}
\usage{
sum_array(x)
}
\arguments{
\item{x}{A numeric vector to sum}
}
\value{
The sum of all elements
}
\description{
Uses loop unrolling (4x) for better CPU pipelining.
}
\examples{
sum_array(1:100)
sum_array(c(1.5, 2.5, 3.0))

}
