% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_array_cpp}
\alias{sum_array_cpp}
\title{Sum a numeric vector using loop unrolling}
\usage{
sum_array_cpp(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The sum of all elements
}
\description{
Sum a numeric vector using loop unrolling
}
