% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateRpackage-package.R
\docType{package}
\name{templateRpackage-package}
\alias{templateRpackage}
\alias{templateRpackage-package}
\title{templateRpackage: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dereckmezquita/template-R-package}
  \item Report bugs at \url{https://github.com/dereckmezquita/template-R-package/issues}
}

}
\author{
\strong{Maintainer}: Dereck Mezquita \email{dereck@mezquita.io} (\href{https://orcid.org/0000-0002-9307-6762}{ORCID})

}
\keyword{internal}
